requirements.txt
requests>=2.25.0
PyJWT>=2.0.0
schedule>=1.1.0
ollama>=0.1.7
pandas>=2.0.0
pandas-ta>=0.3.14
python-dotenv>=1.0.0
python-telegram-bot>=21.0.0
google-genai>=0.1.0
pydantic>=2.0.0
